writing modem values into a html file instead of a list
Before this adjustment there was a modem instance for each modem. These
instances were stored in a list. This list evaluated, a table created
and returnded to the calling website.

Now there is only one modem instance used by all modem request and the
data immediately written into a <tr> to a html file.  This html file
will later on be transfered using  a jquery call.