Adjusting table to display  all values (D2,D3) & fixing table row issue(#5)
The table now displays all modems, d2 and d3. To display both kinds of
modems in the same table, I needed to fill up the empty, "non-bonding"
spaces. I did this by adding empty cells to the value lists.

The tr issue seems to have to do something with the telnet timeout. As
a quickfix I increased it to 1s (try and error). This still needs to be
looked at.