#21 - Threading
Implemented threading using 2 locks, the first one is used to access
the cmts, the second to write the modem object into the html file.

I also added a thread id to identify the treads for debugging purposes.
Will clean this up at a later point

The bootleneck is the cmts. Only 3 to 4 requests per second can be sent
over one telnet session.